SK6812MINI added to ws281X series

+++ components/ws281X/src/DriverWS2811.hpp
@@ -16,11 +16,11 @@ class DriverWS2811 final : public DriverWS281X
 {
 private:
     // It is one half of the time when high speed mode (reset time unchanged)
-    static constexpr uint16_t T0H =   500u; //!< 0 code, high voltage time, ns
-    static constexpr uint16_t T0L =  2000u; //!< 0 code,  low voltage time, ns
-    static constexpr uint16_t T1H =  1200u; //!< 1 code, high voltage time, ns
-    static constexpr uint16_t T1L =  1300u; //!< 1 code,  low voltage time, ns
-    static constexpr uint16_t RES = 50000u; //!< low voltage time, ns
+    static constexpr uint32_t T0H =   500u; //!< 0 code, high voltage time, ns
+    static constexpr uint32_t T0L =  2000u; //!< 0 code,  low voltage time, ns
+    static constexpr uint32_t T1H =  1200u; //!< 1 code, high voltage time, ns
+    static constexpr uint32_t T1L =  1300u; //!< 1 code,  low voltage time, ns
+    static constexpr uint32_t RES = 50000u; //!< low voltage time, ns
 public:
     /**
      * @param mcu  microcontroller unit driver
